+++ src/_cffi_src/openssl/ec.py
@@ -36,6 +36,7 @@ size_t EC_get_builtin_curves(EC_builtin_curve *, size_t);
 
 EC_KEY *EC_KEY_new(void);
 void EC_KEY_free(EC_KEY *);
+int EC_KEY_check_key(const EC_KEY *);
 
 EC_KEY *EC_KEY_new_by_curve_name(int);
 const EC_GROUP *EC_KEY_get0_group(const EC_KEY *);
@@ -45,11 +46,12 @@ const EC_POINT *EC_KEY_get0_public_key(const EC_KEY *);
 int EC_KEY_set_public_key(EC_KEY *, const EC_POINT *);
 void EC_KEY_set_asn1_flag(EC_KEY *, int);
 int EC_KEY_generate_key(EC_KEY *);
-int EC_KEY_set_public_key_affine_coordinates(EC_KEY *, BIGNUM *, BIGNUM *);
 
 EC_POINT *EC_POINT_new(const EC_GROUP *);
 void EC_POINT_free(EC_POINT *);
 
+int EC_POINT_set_affine_coordinates(const EC_GROUP *, EC_POINT *,
+                                    const BIGNUM *, const BIGNUM *, BN_CTX *);
 int EC_POINT_get_affine_coordinates(const EC_GROUP *, const EC_POINT *,
                                     BIGNUM *, BIGNUM *, BN_CTX *);