+++ src/cryptography/hazmat/backends/openssl/ec.py
@@ -72,7 +72,7 @@ def _mark_asn1_named_ec_curve(backend: "Backend", ec_cdata):
     )
 
 
-def _check_key_infinity(backend: "Backend", ec_cdata) -> None:
+def _check_key(backend: "Backend", ec_cdata) -> None:
     point = backend._lib.EC_KEY_get0_public_key(ec_cdata)
     backend.openssl_assert(point != backend._ffi.NULL)
     group = backend._lib.EC_KEY_get0_group(ec_cdata)
@@ -81,6 +81,10 @@ def _check_key_infinity(backend: "Backend", ec_cdata) -> None:
         raise ValueError(
             "Cannot load an EC public key where the point is at infinity"
         )
+    res = backend._lib.EC_KEY_check_key(ec_cdata)
+    if res != 1:
+        backend._consume_errors()
+        raise ValueError("Invalid EC key.")
 
 
 def _sn_to_elliptic_curve(backend: "Backend", sn: str) -> ec.EllipticCurve:
@@ -123,7 +127,14 @@ def _ecdsa_sig_verify(
 
 
 class _EllipticCurvePrivateKey(ec.EllipticCurvePrivateKey):
-    def __init__(self, backend: "Backend", ec_key_cdata, evp_pkey):
+    def __init__(
+        self,
+        backend: "Backend",
+        ec_key_cdata,
+        evp_pkey,
+        *,
+        unsafe_skip_key_validation: bool,
+    ):
         self._backend = backend
         self._ec_key = ec_key_cdata
         self._evp_pkey = evp_pkey
@@ -131,7 +142,9 @@ class _EllipticCurvePrivateKey(ec.EllipticCurvePrivateKey):
         sn = _ec_key_curve_sn(backend, ec_key_cdata)
         self._curve = _sn_to_elliptic_curve(backend, sn)
         _mark_asn1_named_ec_curve(backend, ec_key_cdata)
-        _check_key_infinity(backend, ec_key_cdata)
+        if not unsafe_skip_key_validation:
+            #import pdb;pdb.set_trace()
+            _check_key(backend, ec_key_cdata)
 
     @property
     def curve(self) -> ec.EllipticCurve:
@@ -176,7 +189,14 @@ class _EllipticCurvePrivateKey(ec.EllipticCurvePrivateKey):
 
         evp_pkey = self._backend._ec_cdata_to_evp_pkey(public_ec_key)
 
-        return _EllipticCurvePublicKey(self._backend, public_ec_key, evp_pkey)
+        # We should skip validation here since it was either done on load or
+        # explicitly skipped. Either way we don't do it.
+        return _EllipticCurvePublicKey(
+            self._backend,
+            public_ec_key,
+            evp_pkey,
+            unsafe_skip_key_validation=True,
+        )
 
     def private_numbers(self) -> ec.EllipticCurvePrivateNumbers:
         bn = self._backend._lib.EC_KEY_get0_private_key(self._ec_key)
@@ -215,7 +235,14 @@ class _EllipticCurvePrivateKey(ec.EllipticCurvePrivateKey):
 
 
 class _EllipticCurvePublicKey(ec.EllipticCurvePublicKey):
-    def __init__(self, backend: "Backend", ec_key_cdata, evp_pkey):
+    def __init__(
+        self,
+        backend: "Backend",
+        ec_key_cdata,
+        evp_pkey,
+        *,
+        unsafe_skip_key_validation: bool,
+    ):
         self._backend = backend
         self._ec_key = ec_key_cdata
         self._evp_pkey = evp_pkey
@@ -223,7 +250,9 @@ class _EllipticCurvePublicKey(ec.EllipticCurvePublicKey):
         sn = _ec_key_curve_sn(backend, ec_key_cdata)
         self._curve = _sn_to_elliptic_curve(backend, sn)
         _mark_asn1_named_ec_curve(backend, ec_key_cdata)
-        _check_key_infinity(backend, ec_key_cdata)
+        if not unsafe_skip_key_validation:
+            #import pdb;pdb.set_trace()
+            _check_key(backend, ec_key_cdata)
 
     @property
     def curve(self) -> ec.EllipticCurve: